Able to book slot and identify user to keep in waiting queue

+++ Entities/GamesScheduling/EmployeeCycleStats.cs
@@ -22,6 +22,6 @@ namespace HRMS_Backend.Entities.GamesScheduling
         public GameCycle GameCycle { get; set; }
 
         [Required]
-        public byte GamePlayed { get; set; }
+        public int GamePlayed { get; set; }
     }
 }
